requirements.txt
playwright==1.44.0
anthropic>=0.30.0
python-dotenv>=1.0.0
jinja2>=3.1.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
pdfplumber>=0.10.0
